+++ src/scripts/process/tasks/modals/displayInstallModals.tab.js
@@ -1,7 +1,7 @@
-import {appendModalContainer} from "../../units/page/modals/appendModalContainer.tab";
-import ModalUpdate from "../../units/page/modals/types/ModalUpdate.tab";
-import ModalMessage from "../../units/page/modals/types/ModalMessage.tab";
-import {getStorageData} from "../../units/utils/chromeStorage.all";
+import {appendModalContainer} from "../../../units/page/modals/appendModalContainer.tab";
+import ModalUpdate from "../../../units/page/modals/types/ModalUpdate.tab";
+import ModalMessage from "../../../units/page/modals/types/ModalMessage.tab";
+import {getStorageData} from "../../../units/utils/chromeStorage.all";
 
 export async function launchModalIntegration(domain) {
   // Create modal container
