+++ src/scripts/process/tasks/scraper/scrapping.tab.js
@@ -1,8 +1,8 @@
-import {logWelcome} from "../../units/utils/consoleMessages.all";
-import appInfos from "../../../infos.json";
-import {defineExtractor} from "../../dispatchers/defineExtractor";
-import {defineExportMethod} from "../../dispatchers/defineExportMethod";
-import {updateClickIconCount} from "../../units/internals/icon/clickIconCountContext.bg";
+import {logWelcome} from "../../../units/utils/consoleMessages.all";
+import appInfos from "../../../../infos.json";
+import {defineExtractor} from "../../../dispatchers/defineExtractor";
+import {defineExportMethod} from "../../../dispatchers/defineExportMethod";
+import {updateClickIconCount} from "../../../units/internals/icon/clickIconCountContext.bg";
 
 /**
  * @description - Launch the export process
